chore(git-indexer): drop the SZZ blame pass
fix_events.inducing_shas_json was written on every code fix and read by
nothing. The three surfaces that would have read it are all cut: no UI or
MCP payload names an inducing commit, and the AI-vs-human provenance join
was measured and abandoned. Blame was costing real time to fill a column
with no consumer.

Measured by stubbing the tracer in place and re-running the same git phase,
so the delta is blame alone:

  repowise  46.11s -> 36.98s   -9.13s
  zod       10.31s ->  4.44s   -5.87s
  flask      2.62s ->  2.00s   -0.62s
  django    70.55s -> 72.87s    ~0     (its window holds almost no fixes)

Fix events themselves are untouched: shape, old-side ranges, changed LOC
and committed_at still land per fix-commit x file, which is what the
bug-magnet rollups, per-symbol counts and every shipped surface read.

inducing_shas_json stays on the table, always "[]". Dropping it is a SQLite
table rebuild for two bytes a row, and no migration in this tree drops a
column forward. It is commented as dead so nobody reads the empty list as
"traced and found nothing".

Also renames GitIndexSummary.fix_events_traced to fix_events_built, since
the flag now means the row builder ran rather than the tracer.
